// src/core/enums/core_enums.ts
export enum CoreEnum {
  'RETRIEVAL_REFERENCE_NUMBER' = 'KONGAP',
  'AUTHORIZATION_IDENTIFICATION_RESPONSE' = '000000',
  'CARD_ACCEPTOR_TERMINAL_INTERFACE' = 'KONG    ',
  'CARD_ACCEPTOR_IDENTIFICATION_CODE' = '109839260000000',
  'CURRENCY_CODE_TRANSACTION' = '404',
  'POINT_OF_SERVICE_ENTRY_MODE' = '000',
}

export enum ProcessingCode {
  BALANCE_ENQUIRY = '310000',
  INTERNAL_TRANSFER = '957001',
  FOREX = '930000',
  ACCOUNT_STATEMENT = '922000',
  STANDING_ORDER = '921000',
  STANDING_ORDER_EXTERNAL = '920000',
  CANCEL_STANDING_ORDER = '924000',
  ACCOUNT_CREATION_ONE = '950000',
  ACCOUNT_CREATION_TWO = '961000',
  RTGS = '880001',
}

export enum EnvironmentEnum {
  PRODUCTION = 'PRODUCTION',
  SANDBOX = 'SANDBOX',
}